GIE-228: Support leveraging Perses dashboards
This commit enables listing Perses dashboards from a cluster, getting a
specific one, as well as returning out of the box dashboards that we
ship with OpenShift platform.

The idea is to for the LLM to first look at Out of the box dashboards,
to see if it can answer a user's question using panels from those. And
if not search wider for any custom dashboards a user might have.

We also allow users (and ourselves!) to specify an LLM-friendly
description of PersesDashboard objects using an annotation
operator.perses.dev/mcp-help! This would help the LLM accurately filter
and select dashboards that match a user's query.

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>

Translate list call in handler

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>

pkg/*: leverage existing dashboards and support UI rendering

The MCP server can now find existing Perses dashboards in the cluster
and narrow them down to panels that, upon visualization, help answer the
user's concerns better. It is assumed that since the queries in these
panels have been through human reviews and used in production, they do
not need to be subject to the same validation and guardrails as we for
the queries that the LLM comes up with on-the-fly.

Upon the LLM sharing it's intent to render one or more panels after
going through the existing dashboards, the MCP server will expose
payloads representing panels in the same structure as found in
jhadvig/genie-plugin, so the UI side has no problems graphing these out.

pkg/{mcp,perses}: add tests for added handler parameters

tests,hack: add e2e tests for added tools

pkg: resolve lint issues

hack: resolve Perses CRD issues

pkg: resolve naming conflicts with layout-manager