Add tools for Perses dashboard
This commit enables listing Perses dashboards from a cluster, getting a
specific one, as well as returning out of the box dashboards that we
ship with OpenShift platform.

The idea is to for the LLM to first look at Out of the box dashboards,
to see if it can answer a user's question using panels from those. And
if not search wider for any custom dashboards a user might have.

We also allow users (and ourselves!) to specify an LLM-friendly
description of PersesDashboard objects using an annotation
operator.perses.dev/mcp-help! This would help the LLM accurately filter
and select dashboards that match a user's query.

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>